+++ CHANGES.rst
@@ -11,6 +11,7 @@ Changelog
 - Bugfix: Allow print-listing view on listings within a development.
 - Check for valid development listings.
 - Remove Google Maps API from portal_javascripts.
+- Google Maps now uses configured API key.
 
 
 0.4 (2016-05-20)